ethercat: merge duplicate ALStatusCode definitions
FaultDetection.hpp declared enum class EtherCAT::ALStatusCode while
Reset.hpp declared namespace EtherCAT::ALStatusCode with its own
uint16_t constants — a same-scope collision like the CiA402
OperatingMode/ControlWord cases.

The enum is now canonical; the constants namespace is removed and its
names kept as enum aliases so existing call sites still compile.

While merging, corrected the enum's DC/mailbox values to ETG.1000:
0x0031 InvalidDCLatchConfig, 0x0032 PLLError, 0x0033 DCSyncIOError,
0x0034 DCSyncTimeoutError, 0x0035 InvalidDCCycleTime, 0x0036/37
InvalidDCSync0/1Time; mailbox AoE/EoE/CoE/FoE/SoE at 0x0041-0x0045,
VoE at 0x004F (EoE/CoE/FoE/SoE were previously off-by-one, and the
removed namespace had mailbox codes overlapping the DC range).

Also normalize class/struct forward-declaration tags for
GCode::MachineState and klipper::KlippyInstanceConfig.

+++ tests/sii/zero_coverage_header_tests.cpp
@@ -41,7 +41,7 @@ TEST(WaitResult_Header, Helpers) {
 TEST(Types_Header, CompileTimeConstants) {
     // Basic compile-time checks for constants
     EXPECT_EQ(ALControl::StateMask & ALControl::AckError, 0x0000u);
-    EXPECT_GE(ALStatusCode::VendorSpecificStart, 0x8000u);
+    EXPECT_GE(static_cast<uint16_t>(ALStatusCode::VendorSpecificStart), 0x8000u);
 
     // Ensure PacketFilter size is reasonable (sanity compile check)
     EXPECT_LT(sizeof(PacketFilter), 200u);
